search bar moved to seperate component
