+++ src/component/SearchBar/SearchBar.scss
@@ -1,0 +1,15 @@
+.main {
+  &__top {
+    display: flex;
+    padding: 0 5px;
+  }
+  &__search-bar {
+    // flex-grow: 1;
+    background-image: url("../../assets/icons/search.svg");
+    background-position: left;
+    background-repeat: no-repeat;
+    width: 100%;
+    border-radius: 2rem;
+    padding: 0.4rem 1rem;
+  }
+}
